Update vendored dependency
Prior to this commit the Configuration module dependency was vendored
at `v0.8`. This commit updates the vendored dependency to `v1.2` and
uses the `Save-Module` method instead of cloning the repository. This
ensures that the latest available module is vendored and includes all
of the same code / artifacts as the module available on the gallery.