Add support for GitHub Action environment variables
This adds support for a portion of the environment variables exposed
when running in a GitHub Action. Specifically:

$env:BHBuildSystem
$env:BHProjectPath
$env:BHBranchName
$env:BHCommitMessage

GitHub Actions don't currently have the concept of a build number since
they can be invoked from a number of events against a repo that don't
involve code commits. For this reason, $env:BHBuildNumber will take the
fallback of 0.